Correctly initialize SubStmt of CaseStmt
There is now the option to construct a CaseStmt with a sub statement;
previously, the only way to set a CaseStmt's SubStmt field was to call
the setter function, and this function is not called from anywhere in
the codebase.

This new constructor is used by the AST Importer when importing case
statements.

This commit fixes #23.