Don't collect functions from non-compiling files
The function collector shouldn't collect function names from files which
don't compile because this causes further errors in CTU analysis. Namely
the AST dump files can't be created because of the compilation error,
but the external function map file contains symbols which should be in
these. The error occurs when the AST importer tries to import these
function definitions from the non-existing AST dump file.

Fixes #283